almost completed for attendance marked

+++ src/main/java/com/rtt/attendance/StudentAttendanceMarkingRepository.java
@@ -2,5 +2,5 @@ package com.rtt.attendance;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface StudentAttendanceMarkingRepository extends JpaRepository<AttendanceMarkingEntity,Integer> {
+public interface StudentAttendanceMarkingRepository extends JpaRepository<AttendanceMarkingEntity,Long> {
 }
